Move
Moving get-movies and testing-get-movies in js folder
